refactor: adjust file structure
