+++ .changeset/cyan-geckos-exist.md
@@ -1,0 +1,5 @@
+---
+"react-cool-form": patch
+---
+
+refactor: adjust file structure
